requirements.txt
fastapi
httpx
mac_vendor_lookup
pydantic
aiozmq
pythonping
pyzmq
scapy
uvicorn
